+++ packages/language-server/src/protocol.ts
@@ -40,6 +40,7 @@ export interface DialectDetectionParams {
 export interface IndexedDocument {
   readonly uri: string;
   readonly canonicalUri?: string;
+  readonly mkosiWorkingDirectory?: string;
   readonly languageId: DialectId;
   readonly source: string;
   readonly mtime: number;